Trait robotrs::hid::all::AllTriggerTarget

pub trait AllTriggerTarget: Sized {
    type Error;
    type Output;

    // Provided method
    fn all(self) -> AllTrigger<Self> { ... }
}

An extension trait that is implemented on tuples of triggers that have between 2 and 8 items. All error types must be the same for all triggers.

Required Associated Types

type Error

type Output

Provided Methods

fn all(self) -> AllTrigger<Self>

Create a trigger that activates when all triggers are active

Object Safety

This trait is not object safe.