robotrs::hid::any
Trait AnyTriggerTarget

pub trait AnyTriggerTarget: Sized {
    type Error;
    type Output;

    // Provided method
    fn any(self) -> AnyTrigger<Self> { ... }
}

An extension trait that is automatically implemented for tuples Triggers of length 2 to 8 where all the error and output types are the same.

Required Associated Types

type Error

type Output

Provided Methods

fn any(self) -> AnyTrigger<Self>

Turn this tuple into an AnyTrigger

Object Safety

This trait is not object safe.