Trait robotrs::motor::MotorController

pub trait MotorController: ControlSafe {
    type Error;

    // Required methods
    fn set_percent_raw(&mut self, value: f32) -> Result<(), Self::Error>;
    fn set_voltage(&mut self, value: f32) -> Result<(), Self::Error>;

    // Provided methods
    fn set_percent(&mut self, value: f32) -> Result<(), Self::Error> { ... }
    fn set_inverted(&mut self, is_inverted: bool) -> Result<(), Self::Error> { ... }
}

Defines a motor controller. This trait is implemented for both the VictorSPX and SparkMax. This trait is also implemented for any tuple of MotorControllers

Required Associated Types

type Error

The error type for all operations of the motor controller. Also use by SetIdleMode

Required Methods

fn set_percent_raw(&mut self, value: f32) -> Result<(), Self::Error>

This should only be called with a value between -1 and 1

fn set_voltage(&mut self, value: f32) -> Result<(), Self::Error>

Provided Methods

fn set_percent(&mut self, value: f32) -> Result<(), Self::Error>

Set the percent output of the motor. This automatically clams the value between -1 and 1

fn set_inverted(&mut self, is_inverted: bool) -> Result<(), Self::Error>