robotrs::motor
Trait SetIdleMode

pub trait SetIdleMode: MotorController {
    // Required method
    fn set_idle_mode(&mut self, idle_mode: IdleMode) -> Result<(), Self::Error>;
}

Defines a motor contoller that supports setting the idle mode. This is implemented for the VictorSPX and SparkMax. It is also implemented for tuples.

Required Methods

fn set_idle_mode(&mut self, idle_mode: IdleMode) -> Result<(), Self::Error>

Set the idle mode of the motor to a specific IdleMode