Module robotrs::scheduler

Structs

CancellationHandle

A handle to a cancellation scope. This type uses an Rc, so it can be cloned freely

RobotScheduler

Functions

guard

Create a new cancellation scope with the given future

spawn

Panics if not called after the robot is scheduled or during the robot create closure