robotrs::control
Trait ControlSafe

pub trait ControlSafe {
    // Required method
    fn stop(&mut self);
}

Similar to the wpilib MotorSafety class. The easiest way to properly implement this is to simply call the stop method of all motors and actuators used

Required Methods

fn stop(&mut self)

Stop all motors and actuators. This method gets automatically called whenever a control guard goes out of scope.